Add support for hdr textures

+++ LevEngine/src/Platform/D3D11/D3D11Texture.h
@@ -104,7 +104,7 @@ protected:
 
 	uint8_t m_BPP{};
 
-	uint16_t m_Pitch{};
+	uint32_t m_Pitch{};
 
 	bool m_IsTransparent{};
